servers/requirements.txt
# FastAPI 웹 프레임워크
fastapi~=0.123.0
uvicorn~=0.38.0

# AI
sentence-transformers~=5.1.2
hf_xet~=1.2.0
chromadb~=1.3.5

# 데이터 검증
pydantic~=2.12.5

# OpenAI API
openai~=2.8.1
python-dotenv~=1.1.1

# PDF Reader
pypdf~=6.4.0

# 파일 업로드 처리
python-multipart~=0.0.20
